+++ app/src/main/AndroidManifest.xml
@@ -13,18 +13,18 @@
         android:theme="@style/Theme.ATSOPTANDROID"
         tools:targetApi="31">
         <activity
-            android:name=".MyActivity"
+            android:name=".presentation.ui.my.MyActivity"
             android:exported="false"
             android:label="@string/title_activity_my"
             android:theme="@style/Theme.ATSOPTANDROID" />
         <activity
-            android:name=".signup.SignUpActivity"
+            android:name=".presentation.ui.signup.SignUpActivity"
             android:exported="false"
             android:label="@string/title_activity_sign_up"
             android:theme="@style/Theme.ATSOPTANDROID"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".SignInActivity"
+            android:name=".presentation.ui.signin.SignInActivity"
             android:exported="true"
             android:label="@string/title_activity_sign_in"
             android:theme="@style/Theme.ATSOPTANDROID">
