+++ app/src/main/java/org/sopt/at/presentation/ui/signup/IdInputScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.signup
+package org.sopt.at.presentation.ui.signup
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -21,33 +21,29 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import kotlinx.coroutines.launch
-import org.sopt.at.PasswordTextField
 
 @Composable
-fun PwInputScreen(
-    pw: String,
-    onPwChange: (String) -> Unit,
+fun IdInputScreen(
+    id: String,
+    onIdChange: (String) -> Unit,
     onNext: () -> Unit,
     onBack: () -> Unit,
     modifier: Modifier = Modifier
 ) {
-    var showPassword by remember { mutableStateOf(value = false) }
-
-    // pw 유효성 검사: 영문, 숫자, 특수문자(~!@#$%^&*) 조합 8~15자리
-    val isValidPw = pw.matches(Regex("^(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[~!@#\$%^&*])[a-zA-Z0-9~!@#\$%^&*]{8,15}$"))
+    // id 유효성 검사: 영문 소문자 또는 영문 소문자, 숫자 조합 6~12자리
+    val isValidId = id.matches(Regex("^(?=.*[a-z])[a-z0-9]{6,12}$"))
 
     // 스낵바
     val scope = rememberCoroutineScope()
@@ -61,12 +57,12 @@ fun PwInputScreen(
         Column(
             modifier = Modifier
                 .fillMaxSize()
-                .imePadding()
                 .background(color = Color.Black)
                 .padding(innerPadding)
-                .padding(15.dp)
+                .padding(15.dp),
         ) {
             // 뒤로 가기 버튼
+            // 클릭 시 로그인 뷰로 이동
             Box(
                 modifier = Modifier.size(24.dp)
             ) {
@@ -88,32 +84,49 @@ fun PwInputScreen(
             Column(
                 modifier = Modifier
                     .fillMaxSize()
-                    .padding(top = 15.dp),
+                    .padding(top = 15.dp)
+                    .imePadding(),
                 verticalArrangement = Arrangement.SpaceBetween
             ) {
-                // 비밀번호 입력
+                // id 입력
                 Column(
                     horizontalAlignment = Alignment.CenterHorizontally
                 ) {
                     Text(
-                        text = "비밀번호를 입력해주세요.",
+                        text = "아이디를 입력해주세요.",
                         color = Color.White,
                         fontSize = 20.sp
                     )
 
                     Column {
-                        PasswordTextField(
-                            password = pw,
-                            onPasswordChange = onPwChange,
-                            showPassword = showPassword,
-                            onTogglePasswordVisibility = { showPassword = !showPassword },
-                            modifier = Modifier
+                        // 아이디 입력 창
+                        TextField(
+                            value = id,
+                            onValueChange = onIdChange,
+                            placeholder = { Text("아이디") },
+                            modifier = modifier
+                                .fillMaxWidth()
                                 .padding(top = 20.dp)
                                 .border(1.dp, Color.Gray, shape = RoundedCornerShape(5.dp)),
+                            shape = RoundedCornerShape(5.dp),
+                            colors = TextFieldDefaults.colors(
+                                focusedContainerColor = Color(0xFF262626),
+                                unfocusedContainerColor = Color(0xFF262626),
+                                focusedIndicatorColor = Color.Transparent,
+                                unfocusedIndicatorColor = Color.Transparent,
+                                disabledIndicatorColor = Color.Transparent,
+                                cursorColor = Color.White,
+                                focusedPlaceholderColor = Color.Gray,
+                                unfocusedPlaceholderColor = Color.Gray
+                            ),
+                            textStyle = TextStyle(
+                                color = Color.White
+                            ),
+                            singleLine = true
                         )
 
                         Text(
-                            text = "영문, 숫자, 특수문자(~!@#$%^) 조합 8~15자리",
+                            text = "영문 소문자 또는 영문 소문자, 숫자 조합 6~12 자리",
                             color = Color.Gray,
                             fontSize = 12.sp
                         )
@@ -132,13 +145,13 @@ fun PwInputScreen(
                             interactionSource = remember { MutableInteractionSource() },
                             indication = null
                         ) {
-                            // pw가 유효할 경우 로그인 뷰로 이동
-                            if (isValidPw) {
+                            // id가 유효할 경우 회원가입 비밀번호 뷰로 이동
+                            if (isValidId) {
                                 onNext()
                             } else {
-                                // pw가 유효하지 않을 경우 스낵바
+                                // id가 유효 하지 않을 시 스낵바
                                 scope.launch {
-                                    snackbarHostState.showSnackbar("비밀번호가 유효하지 않습니다.")
+                                    snackbarHostState.showSnackbar("ID가 유효하지 않습니다.")
                                 }
                             }
                         },
@@ -150,9 +163,7 @@ fun PwInputScreen(
                         color = Color.LightGray
                     )
                 }
-
             }
-
         }
     }
 }
